# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Market API
  version: 1.0.0
  description: |
    Backend API cho **Smart Market – Chợ thông minh** (Zalo Mini App + CMS).

    ## Xác thực
    - **Mini app**: `POST /auth/zalo-login` với access token Zalo. Môi trường dev bật `ZALO_MOCK=true`
      cho phép token dạng `mock:<zaloId>:<Tên>` (vd `mock:zalo-buyer-1:Người Mua`).
    - **CMS**: `POST /auth/login` với username/email/số điện thoại + mật khẩu.
    - Sau khi đăng nhập, bấm **Authorize** và dán `accessToken` (không cần chữ "Bearer").

    ## Vai trò & phạm vi
    | Role | Phạm vi |
    |---|---|
    | super_admin | Toàn hệ thống |
    | province_admin | Các chợ trong tỉnh được gán (chủ yếu chỉ đọc + escalate phản ánh) |
    | market_manager | CRUD trong các chợ được gán |
    | user | Người mua; nếu `merchant_status=active` và có sạp → dùng được nhóm `/seller/*` |

    ## Quy ước
    - Envelope: `{ success, data, meta? }` hoặc `{ success: false, error: { code, message, details? } }`.
    - Phân trang: `?page=1&limit=20` (limit tối đa 100) → `meta { page, limit, total, totalPages }`.
    - Tiền VND số nguyên. Thời gian ISO 8601. Key JSON camelCase.
    - Tài khoản demo (seed): SA `admin@smartmarket.vn / Admin@12345`; MM `mm.trungtam@smartmarket.vn / Manager@123`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MarketIdQuery:
      name: marketId
      in: query
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: STATE_INVALID }
            message: { type: string, example: Đơn đang ở trạng thái "completed", không thể chuyển sang "confirmed" }
            details: { nullable: true }
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, preparing, ready, completed, cancelled]
...
  responses:
    Ok:
      description: Thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: {}
    Paginated:
      description: Danh sách có phân trang
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: { type: array, items: {} }
              meta:
                type: object
                properties:
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  total: { type: integer, example: 135 }
                  totalPages: { type: integer, example: 7 }
    Err:
      description: Lỗi (400/401/403/404/409/429 — xem `error.code`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security: []
paths:
...
  /orders:
    post:
      tags: [Orders (Buyer)]
      summary: Checkout — tách đơn theo sạp, nhóm bằng checkoutGroupId
      description: |
        Giỏ hàng ở client. `expectedUnitPrice` = finalPrice client đang thấy —
        lệch giá server trả `409 PRICE_CHANGED` kèm giá mới. Thiếu kho → `409 STOCK_NOT_ENOUGH`.
        MVP chỉ hỗ trợ `receiveType=pickup`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [marketId, receiverName, receiverPhone, items]
              properties:
                marketId: { type: string, format: uuid }
                receiveType: { type: string, enum: [pickup, delivery], default: pickup }
                receiverName: { type: string }
                receiverPhone: { type: string, example: '0922000001' }
                note: { type: string, nullable: true }
                items:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: object
                    required: [productId, quantity, expectedUnitPrice]
                    properties:
                      productId: { type: string, format: uuid }
                      quantity: { type: number, example: 2 }
                      expectedUnitPrice: { type: integer, example: 22500 }
      responses:
        '201':
          description: Danh sách đơn đã tạo
          content:
            application/json:
              example:
                success: true
                data:
                  checkoutGroupId: 5f3c...
                  orders: [{ id: '...', code: SM260711-X7K2P, totalAmount: 55000, status: pending }]
        default: { $ref: '#/components/responses/Err' }
    get:
      tags: [Orders (Buyer)]
      summary: Đơn hàng của tôi
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
  /orders/{id}:
    get:
      tags: [Orders (Buyer)]
      summary: Chi tiết đơn (kèm items, timeline)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /orders/{id}/cancel:
    post:
      tags: [Orders (Buyer)]
      summary: Buyer hủy đơn (khi pending/confirmed) — tự hoàn kho
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, nullable: true }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }

  # ================= REVIEWS =================
...
  /seller/orders:
    get:
      tags: [Seller]
      summary: Đơn hàng của sạp
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
  /seller/orders/{id}:
    get:
      tags: [Seller]
      summary: Chi tiết đơn của sạp
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /seller/orders/{id}/status:
    patch:
      tags: [Seller]
      summary: 'Chuyển trạng thái tuần tự: pending→confirmed→preparing→ready→completed'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [confirmed, preparing, ready, completed] }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /seller/orders/{id}/cancel:
    post:
      tags: [Seller]
      summary: Seller hủy đơn (bắt buộc lý do) — tự hoàn kho
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }

  # ================= ADMIN - DASHBOARD =================
...
  /admin/orders:
    get:
      tags: [Admin - Orders]
      summary: Danh sách đơn theo scope
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MarketIdQuery'
        - { name: stallId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
  /admin/orders/{id}:
    get:
      tags: [Admin - Orders]
      summary: Chi tiết đơn
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /admin/orders/{id}/cancel:
    post:
      tags: [Admin - Orders]
      summary: Admin hủy đơn xử lý sự cố (bắt buộc lý do, hoàn kho, báo 2 phía)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }

  # ================= ADMIN - COMPLAINTS =================
...